+++ CHANGELOG.md
@@ -12,6 +12,8 @@ FlatLaf Change Log
   scrollable.
 - Added system property `flatlaf.useNativeLibrary` to allow disabling loading of
   FlatLaf native library. (issue #674)
+- IntelliJ Themes: Reduced memory footprint and improved setup speed by ignoring
+  IntelliJ UI properties that are not used in FlatLaf.
 
 #### Fixed bugs